% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{abs1_}
\alias{abs1_}
\title{abs() element-wise absolute value demo (real and complex)}
\usage{
abs1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): sum of abs(real matrix) and abs(complex matrix)
}
\description{
abs() element-wise absolute value demo (real and complex)
}
