% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{affmul1_}
\alias{affmul1_}
\title{affmul() affine matrix-vector multiplication demo}
\usage{
affmul1_(n)
}

\arguments{
\item{n}{integer scalar; row dimension}
}

\value{
numeric vector (N): result of `A[1:N, 1:N] \%*\% B + A[, N+1]`
}

\description{
affmul() affine matrix-vector multiplication demo
}

