% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{all1_}
\alias{all1_}
\title{all() all-elements-nonzero check demo}
\usage{
all1_(n)
}
\arguments{
\item{n}{integer scalar; container dimension}
}
\value{
logical vector of four all() check results
}
\description{
all() all-elements-nonzero check demo
}
