% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{attr1_}
\alias{attr1_}
\title{Matrix dimension attributes}
\usage{
attr1_(a)
}
\arguments{
\item{a}{numeric matrix}
}
\value{
named integer vector with elements \code{n_rows}, \code{n_cols}, and \code{n_elem}
}
\description{
Matrix dimension attributes
}
