% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{chol1_}
\alias{chol1_}
\title{chol() Cholesky decomposition with pivoting demo}
\usage{
chol1_(x, layout, output)
}
\arguments{
\item{x}{numeric positive semi-definite matrix (N*N)}

\item{layout}{string scalar; \code{"upper"} or \code{"lower"}}

\item{output}{string scalar; \code{"matrix"} for pivot matrix, \code{"vector"} for
pivot vector}
}
\value{
list with success flag and triangular factor matrix
}
\description{
chol() Cholesky decomposition with pivoting demo
}
