% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{chol_mat}
\alias{chol_mat}
\title{Cholesky decomposition}
\usage{
chol_mat(x, type)
}
\arguments{
\item{x}{symmetric positive definite numeric matrix (N*N)}

\item{type}{decomposition type: \code{"upper"} for upper triangular or \code{"lower"}
for lower triangular}
}
\value{
triangular factor matrix (N*N)
}
\description{
Cholesky decomposition
}
