% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{clamp1_}
\alias{clamp1_}
\title{clamp() value-clamping demo}
\usage{
clamp1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N) with values clamped to the interval 0.2 to 0.8
}

\description{
clamp() value-clamping demo
}

