% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{clean1_}
\alias{clean1_}
\title{clean() small-value zeroing demo}
\usage{
clean1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N) with near-zero elements set to zero
}

\description{
clean() small-value zeroing demo
}

