% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{colptr1_}
\alias{colptr1_}
\title{colptr() column memory pointer access demo}
\usage{
colptr1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N) with first column elements increased by 123
}
\description{
colptr() column memory pointer access demo
}
