% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{compatibility1_}
\alias{compatibility1_}
\title{front() and back() compatibility demo}
\usage{
compatibility1_(n)
}

\arguments{
\item{n}{integer scalar; vector length}
}

\value{
named numeric vector with `front` and `back` elements
}

\description{
front() and back() compatibility demo
}

