% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cond1_}
\alias{cond1_}
\title{cond() matrix condition number demo}
\usage{
cond1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric scalar: condition number of the N*N identity matrix (1)
}
\description{
cond() matrix condition number demo
}
