% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{conj1_}
\alias{conj1_}
\title{conj() complex conjugate demo}
\usage{
conj1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
complex matrix (N*N): complex conjugate of a random complex matrix
}
\description{
conj() complex conjugate demo
}
