% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{conv2_}
\alias{conv2_}
\title{conv2() 2D convolution demo (full and same modes)}
\usage{
conv2_(x, y)
}

\arguments{
\item{x}{numeric matrix (kernel)}

\item{y}{numeric matrix (signal)}
}

\value{
list with full-size and same-size 2D convolution results
}

\description{
conv2() 2D convolution demo (full and same modes)
}

