% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{copy_size1_}
\alias{copy_size1_}
\title{copy_size() demo}
\usage{
copy_size1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
named integer vector with \code{n_rows} and \code{n_cols} of the copied matrix
}
\description{
copy_size() demo
}
