% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{cor1_}
\alias{cor1_}
\title{cor() correlation matrix demo (sample and population normalization)}
\usage{
cor1_(X, Y)
}

\arguments{
\item{X}{numeric matrix (first variable)}

\item{Y}{numeric matrix (second variable)}
}

\value{
list with two correlation matrices
}

\description{
cor() correlation matrix demo (sample and population normalization)
}

