% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cov1_}
\alias{cov1_}
\title{cov() covariance matrix demo (sample and population normalization)}
\usage{
cov1_(X, Y)
}
\arguments{
\item{X}{numeric matrix (first variable)}

\item{Y}{numeric matrix (second variable)}
}
\value{
list with two covariance matrices
}
\description{
cov() covariance matrix demo (sample and population normalization)
}
