% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{diagvec1_}
\alias{diagvec1_}
\title{diagvec() diagonal extraction demo (super and sub-diagonal)}
\usage{
diagvec1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric vector (N-1): sum of first upper and lower diagonals
}
\description{
diagvec() diagonal extraction demo (super and sub-diagonal)
}
