% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{diff1_}
\alias{diff1_}
\title{diff() first-order differences demo}
\usage{
diff1_(n)
}

\arguments{
\item{n}{integer scalar; vector length}
}

\value{
numeric matrix (N*2): original vector and its first differences
}

\description{
diff() first-order differences demo
}

