% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{each_col1_}
\alias{each_col1_}
\title{each_col() broadcast demo}
\usage{
each_col1_(n)
}

\arguments{
\item{n}{integer scalar; number of rows}
}

\value{
numeric matrix (N*(N+1)): result of column-wise broadcasts and
  lambda operations
}

\description{
each_col() broadcast demo
}

