% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{each_row1_}
\alias{each_row1_}
\title{each_row() broadcast demo}
\usage{
each_row1_(n)
}
\arguments{
\item{n}{integer scalar; number of columns}
}
\value{
numeric matrix ((N+1)*N): result of row-wise broadcasts and
lambda operations
}
\description{
each_row() broadcast demo
}
