% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{eig_gen1_}
\alias{eig_gen1_}
\title{eig_gen() general eigendecomposition demo with optional balancing}
\usage{
eig_gen1_(x, balance)
}

\arguments{
\item{x}{numeric matrix (N*N)}

\item{balance}{string scalar; `"balance"` or `"nobalance"`}
}

\value{
list with success flag, complex eigenvalue vector, and complex
  eigenvector matrix
}

\description{
eig_gen() general eigendecomposition demo with optional balancing
}

