% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{eig_gen2_}
\alias{eig_gen2_}
\title{eigs_gen() sparse general eigendecomposition demo}
\usage{
eig_gen2_(x, method, k)
}

\arguments{
\item{x}{numeric matrix (used as a sparse matrix)}

\item{method}{string scalar; `"lm"` (largest magnitude) or `"sm"`}

\item{k}{integer scalar; number of eigenvalues to compute}
}

\value{
list with success flag, complex eigenvalue vector, and complex
  eigenvector matrix
}

\description{
eigs_gen() sparse general eigendecomposition demo
}

