% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{eig_pair1_}
\alias{eig_pair1_}
\title{eig_pair() generalized eigendecomposition (A\emph{x = lambda}B*x) demo}
\usage{
eig_pair1_(a, b)
}
\arguments{
\item{a}{numeric matrix (N*N)}

\item{b}{numeric matrix (N*N)}
}
\value{
list with success flag, complex eigenvalue vector, and complex
eigenvector matrix
}
\description{
eig_pair() generalized eigendecomposition (A\emph{x = lambda}B*x) demo
}
