% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{eig_sym1_}
\alias{eig_sym1_}
\title{eig_sym() symmetric eigendecomposition demo (two forms)}
\usage{
eig_sym1_(x, method)
}
\arguments{
\item{x}{symmetric numeric matrix (N*N)}

\item{method}{string scalar; \code{"dc"} (divide-and-conquer) or \code{"std"}}
}
\value{
list with success flag, eigenvalue vector, and eigenvector matrix
}
\description{
eig_sym() symmetric eigendecomposition demo (two forms)
}
