% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{eigen_gen_dbl_complex_wrapper}
\alias{eigen_gen_dbl_complex_wrapper}
\title{Eigenvalues of a general matrix as a complex vector}
\usage{
eigen_gen_dbl_complex_wrapper(x)
}
\arguments{
\item{x}{numeric matrix (N*N)}
}
\value{
complex eigenvalue vector (N)
}
\description{
Eigenvalues of a general matrix as a complex vector
}
