% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{eigen_gen_mat}
\alias{eigen_gen_mat}
\title{Eigenvalues of a general matrix (real and imaginary parts)}
\usage{
eigen_gen_mat(x)
}

\arguments{
\item{x}{numeric matrix (N*N)}
}

\value{
list with elements `real` and `imag` (each N*1 matrix)
}

\description{
Eigenvalues of a general matrix (real and imaginary parts)
}

