% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{eigen_gen_mat_complex_wrapper}
\alias{eigen_gen_mat_complex_wrapper}
\title{Eigenvalues of a general matrix as a complex matrix}
\usage{
eigen_gen_mat_complex_wrapper(x)
}

\arguments{
\item{x}{numeric matrix (N*N)}
}

\value{
complex eigenvalue matrix (N*1)
}

\description{
Eigenvalues of a general matrix as a complex matrix
}

