% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{eigen_gen_no_wrapper}
\alias{eigen_gen_no_wrapper}
\title{Eigenvalues of a general matrix without wrapper helpers}
\usage{
eigen_gen_no_wrapper(x)
}
\arguments{
\item{x}{numeric matrix (N*N)}
}
\value{
list with elements \code{real} and \code{imag} (each N*1 matrix)
}
\description{
Eigenvalues of a general matrix without wrapper helpers
}
