% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{fft1_}
\alias{fft1_}
\title{fft() and ifft() 1D fast Fourier transform demo}
\usage{
fft1_(x)
}

\arguments{
\item{x}{numeric vector}
}

\value{
list with FFT result (real/imag) and inverse FFT result (real/imag)
}

\description{
fft() and ifft() 1D fast Fourier transform demo
}

