% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{field_imat_demo_}
\alias{field_imat_demo_}
\title{Round-trip list of integer matrices via field<imat>}
\usage{
field_imat_demo_(mats)
}

\arguments{
\item{mats}{list of integer matrices}
}

\value{
list of integer matrices identical to input
}

\description{
Round-trip list of integer matrices via field<imat>
}

