% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{field_mat_demo_}
\alias{field_mat_demo_}
\title{Normalize each matrix in a list via field<mat>}
\usage{
field_mat_demo_(mats)
}

\arguments{
\item{mats}{list of numeric matrices}
}

\value{
list of normalized numeric matrices
}

\description{
Normalize each matrix in a list via field<mat>
}

