% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{find1_}
\alias{find1_}
\title{find() conditional index search demo}
\usage{
find1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with three integer vectors of indices satisfying different
conditions
}
\description{
find() conditional index search demo
}
