% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{find_finite1_}
\alias{find_finite1_}
\title{find_finite() finite element index search demo}
\usage{
find_finite1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
integer vector: linear indices of finite elements
}

\description{
find_finite() finite element index search demo
}

