% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{find_nan1_}
\alias{find_nan1_}
\title{find_nan() NaN element index search demo}
\usage{
find_nan1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
integer vector: linear indices of NaN elements
}

\description{
find_nan() NaN element index search demo
}

