% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{find_nonfinite1_}
\alias{find_nonfinite1_}
\title{find_nonfinite() non-finite element index search demo}
\usage{
find_nonfinite1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
integer vector: linear indices of non-finite (Inf/NaN) elements
}

\description{
find_nonfinite() non-finite element index search demo
}

