% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{find_unique1_}
\alias{find_unique1_}
\title{find_unique() unique element index search demo}
\usage{
find_unique1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
integer vector: linear indices of unique elements
}

\description{
find_unique() unique element index search demo
}

