% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{flip1_}
\alias{flip1_}
\title{fliplr() / flipud() matrix flip demo}
\usage{
flip1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with original matrix, left-right flip, and up-down flip
}
\description{
fliplr() / flipud() matrix flip demo
}
