% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{for_each1_}
\alias{for_each1_}
\title{for_each() iteration demo on dense, sparse, and field containers}
\usage{
for_each1_(n)
}

\arguments{
\item{n}{integer scalar; container dimension}
}

\value{
numeric matrix (N*N): sum of modified dense matrix, sparse matrix,
  and two zero field matrices
}

\description{
for_each() iteration demo on dense, sparse, and field containers
}

