% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{has_nan1_}
\alias{has_nan1_}
\title{has_nan() NaN check demo (dense, cube, sparse)}
\usage{
has_nan1_(n)
}

\arguments{
\item{n}{integer scalar; container dimension}
}

\value{
named logical vector of three has-NaN predicate results
}

\description{
has_nan() NaN check demo (dense, cube, sparse)
}

