% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{hist1_}
\alias{hist1_}
\title{hist() histogram with fixed and custom bin edges demo}
\usage{
hist1_(n)
}
\arguments{
\item{n}{integer scalar; number of random values}
}
\value{
list with two integer count vectors
}
\description{
hist() histogram with fixed and custom bin edges demo
}
