% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{histc1_}
\alias{histc1_}
\title{histc() cumulative histogram demo}
\usage{
histc1_(n)
}

\arguments{
\item{n}{integer scalar; number of random values}
}

\value{
integer vector: bin counts using custom edges
}

\description{
histc() cumulative histogram demo
}

