% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{imag1_}
\alias{imag1_}
\title{imag() and real() complex part extraction demo}
\usage{
imag1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with `imag` and `real` parts of a random complex matrix
}

\description{
imag() and real() complex part extraction demo
}

