% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{imbue1_}
\alias{imbue1_}
\title{imbue fill via C++ RNG demo}
\usage{
imbue1_(a)
}
\arguments{
\item{a}{numeric matrix}
}
\value{
numeric matrix: \code{a} plus a matrix filled via Mersenne-twister uniform
distribution
}
\description{
imbue fill via C++ RNG demo
}
