% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{imbue2_}
\alias{imbue2_}
\title{imbue fill via R's RNG demo}
\usage{
imbue2_(a)
}

\arguments{
\item{a}{numeric matrix}
}

\value{
numeric matrix: `a` plus a matrix filled via `unif_rand()`
}

\description{
imbue fill via R's RNG demo
}

