% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{implicit_conversion_double}
\alias{implicit_conversion_double}
\title{Round-trip R matrix via arma::mat (double precision)}
\usage{
implicit_conversion_double(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
numeric matrix identical to input
}
\description{
Round-trip R matrix via arma::mat (double precision)
}
