% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{implicit_conversion_float}
\alias{implicit_conversion_float}
\title{Round-trip R matrix via arma::fmat (single precision)}
\usage{
implicit_conversion_float(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
numeric matrix (values may differ slightly due to float precision)
}
\description{
Round-trip R matrix via arma::fmat (single precision)
}
