% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{implicit_conversion_uint}
\alias{implicit_conversion_uint}
\title{Round-trip R matrix via arma::umat (unsigned integer)}
\usage{
implicit_conversion_uint(x)
}

\arguments{
\item{x}{integer or numeric matrix}
}

\value{
integer matrix (values truncated to unsigned range)
}

\description{
Round-trip R matrix via arma::umat (unsigned integer)
}

