% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{in_range1_}
\alias{in_range1_}
\title{in_range() index validity check demo}
\usage{
in_range1_(n)
}
\arguments{
\item{n}{integer scalar; matrix row dimension}
}
\value{
named logical vector indicating which row/column pairs are in range
}
\description{
in_range() index validity check demo
}
