% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{index_maxmin1_}
\alias{index_maxmin1_}
\title{index_max() and index_min() demo}
\usage{
index_maxmin1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
named numeric vector: linear indices of max/min and four corner
  element values
}

\description{
index_max() and index_min() demo
}

