% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{initialization1_}
\alias{initialization1_}
\title{Brace-initializer construction demo}
\usage{
initialization1_(a)
}
\arguments{
\item{a}{numeric matrix}
}
\value{
numeric matrix: \code{a} plus a 2*2 brace-initialized matrix with diagonal
shifted by a brace-initialized vector
}
\description{
Brace-initializer construction demo
}
