% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{inplace_strans1_}
\alias{inplace_strans1_}
\title{inplace_strans() in-place conjugate transpose demo}
\usage{
inplace_strans1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
complex matrix (N*N): complex matrix after in-place conjugate
  transpose
}

\description{
inplace_strans() in-place conjugate transpose demo
}

