% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{insert_rows1_}
\alias{insert_rows1_}
\title{insert_rows() row insertion demo}
\usage{
insert_rows1_(n)
}

\arguments{
\item{n}{integer scalar; base matrix dimension}
}

\value{
numeric matrix (3N-1)*(3N-1) with rows inserted
}

\description{
insert_rows() row insertion demo
}

