% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{insert_slices1_}
\alias{insert_slices1_}
\title{insert_slices() cube insertion demo}
\usage{
insert_slices1_(n)
}
\arguments{
\item{n}{integer scalar; cube dimension}
}
\value{
numeric matrix (N*N): element-wise sum over all slices of two
modified cubes
}
\description{
insert_slices() cube insertion demo
}
