% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{integer_cube_roundtrip_}
\alias{integer_cube_roundtrip_}
\title{Round-trip list of matrices via Cube\if{html}{\out{<int>}}}
\usage{
integer_cube_roundtrip_(slices)
}
\arguments{
\item{slices}{list of integer matrices}
}
\value{
list of integer matrices identical to input
}
\description{
Round-trip list of matrices via Cube\if{html}{\out{<int>}}
}
