% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{interp1_}
\alias{interp1_}
\title{interp1() 1D interpolation demo (linear and nearest-neighbor)}
\usage{
interp1_(n)
}

\arguments{
\item{n}{integer scalar; number of data points}
}

\value{
numeric vector (2N): interpolated values at finer grid
}

\description{
interp1() 1D interpolation demo (linear and nearest-neighbor)
}

