% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{intersect1_}
\alias{intersect1_}
\title{intersect() sorted-set intersection demo}
\usage{
intersect1_(n)
}

\arguments{
\item{n}{integer scalar; sequence length}
}

\value{
integer vector: intersection of `n:1` and `2:(n+1)`
}

\description{
intersect() sorted-set intersection demo
}

