% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{inv_sympd1_}
\alias{inv_sympd1_}
\title{inv_sympd() symmetric positive definite matrix inversion demo}
\usage{
inv_sympd1_(a)
}

\arguments{
\item{a}{numeric symmetric positive definite matrix (N*N)}
}

\value{
list with success flag and inverse matrix
}

\description{
inv_sympd() symmetric positive definite matrix inversion demo
}

