% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{inverse1_}
\alias{inverse1_}
\title{Matrix inverse and linear solve demo}
\usage{
inverse1_(a, b)
}
\arguments{
\item{a}{numeric square matrix (N*N)}

\item{b}{numeric vector (N)}
}
\value{
numeric vector: \code{inv(a) \%*\% b}
}
\description{
Matrix inverse and linear solve demo
}
