% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{is_hermitian1_}
\alias{is_hermitian1_}
\title{is_hermitian() Hermitian check demo}
\usage{
is_hermitian1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
named logical vector of three Hermitian predicate results
}

\description{
is_hermitian() Hermitian check demo
}

