% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{is_sorted1_}
\alias{is_sorted1_}
\title{is_sorted() sortedness check demo}
\usage{
is_sorted1_(n)
}

\arguments{
\item{n}{integer scalar; vector length}
}

\value{
named logical vector of four sortedness check results
}

\description{
is_sorted() sortedness check demo
}

