% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{is_symmetric1_}
\alias{is_symmetric1_}
\title{is_symmetric() symmetry check demo}
\usage{
is_symmetric1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
named logical vector of three symmetry predicate results
}

\description{
is_symmetric() symmetry check demo
}

