% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{is_sympd1_}
\alias{is_sympd1_}
\title{is_sympd() symmetric positive definiteness check demo}
\usage{
is_sympd1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
named logical vector of three sym-pos-def predicate results
}

\description{
is_sympd() symmetric positive definiteness check demo
}

