% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{is_vec1_}
\alias{is_vec1_}
\title{is_vec() / is_colvec() / is_rowvec() predicate demo}
\usage{
is_vec1_(n)
}
\arguments{
\item{n}{integer scalar; vector length}
}
\value{
named logical vector of five vector-shape predicate results
}
\description{
is_vec() / is_colvec() / is_rowvec() predicate demo
}
