% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{iterators1_}
\alias{iterators1_}
\title{Element and column iterator demo on a matrix}
\usage{
iterators1_(n)
}
\arguments{
\item{n}{integer scalar; matrix row dimension}
}
\value{
numeric matrix (N*(N+1)) after iterating all elements and column 1-N
}
\description{
Element and column iterator demo on a matrix
}
