% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{iterators3_}
\alias{iterators3_}
\title{Sparse matrix element iterator demo}
\usage{
iterators3_(n)
}
\arguments{
\item{n}{integer scalar; matrix row dimension}
}
\value{
numeric matrix (N*(2N)): sparse matrix after incrementing all
non-zero elements by 123
}
\description{
Sparse matrix element iterator demo
}
