% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{join_cubes1_}
\alias{join_cubes1_}
\title{join_slices() cube concatenation along slice dimension demo}
\usage{
join_cubes1_(n)
}
\arguments{
\item{n}{integer scalar; cube row/column dimension}
}
\value{
list of 7 slice matrices from a cube formed by joining two cubes
}
\description{
join_slices() cube concatenation along slice dimension demo
}
