% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{join_rows1_}
\alias{join_rows1_}
\title{join_rows() / join_cols() matrix concatenation demo}
\usage{
join_rows1_(n)
}
\arguments{
\item{n}{integer scalar; matrix row dimension}
}
\value{
list with three N*1 input matrices, their horizontal concatenation,
and their vertical concatenation
}
\description{
join_rows() / join_cols() matrix concatenation demo
}
