% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{kron1_}
\alias{kron1_}
\title{kron() Kronecker product demo}
\usage{
kron1_(n)
}

\arguments{
\item{n}{integer scalar; base matrix dimension}
}

\value{
numeric matrix (N*(N+1) * (N+1)*N): Kronecker product
}

\description{
kron() Kronecker product demo
}

