% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{linspace1_}
\alias{linspace1_}
\title{linspace() linearly spaced vector demo}
\usage{
linspace1_(n)
}
\arguments{
\item{n}{integer scalar; number of points}
}
\value{
numeric vector (N): sum of a column and a row linspace vector
}
\description{
linspace() linearly spaced vector demo
}
