% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{log_det1_}
\alias{log_det1_}
\title{log_det() log-determinant demo (complex and real/sign forms)}
\usage{
log_det1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with complex form result (real/imag) and real form result
  (val/sign/ok)
}

\description{
log_det() log-determinant demo (complex and real/sign forms)
}

