% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{logmat_sympd1_}
\alias{logmat_sympd1_}
\title{logmat_sympd() logarithm of symmetric positive definite matrix demo}
\usage{
logmat_sympd1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with two elements: the result matrix and a sub-list with result
  matrix plus success flag
}

\description{
logmat_sympd() logarithm of symmetric positive definite matrix demo
}

