% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{mat_out_demo_}
\alias{mat_out_demo_}
\title{Armadillo-to-R matrix conversion demo: t(A) * A}
\usage{
mat_out_demo_(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
numeric matrix: crossproduct t(x) * x
}
\description{
Armadillo-to-R matrix conversion demo: t(A) * A
}
