% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{matrix_add_}
\alias{matrix_add_}
\title{Matrix addition using cpp4r::as_sexp()}
\usage{
matrix_add_(a, b)
}

\arguments{
\item{a}{numeric matrix}

\item{b}{numeric matrix}
}

\value{
numeric matrix: element-wise sum of `a` and `b`
}

\description{
Matrix addition using cpp4r::as_sexp()
}

