% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{max1_}
\alias{max1_}
\title{max() column-wise, row-wise, and element-wise maximum demo}
\usage{
max1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with column max, row max, column max (dim=0), and element-wise
max of two matrices
}
\description{
max() column-wise, row-wise, and element-wise maximum demo
}
