% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{maxmin1_}
\alias{maxmin1_}
\title{max() and min() of a matrix demo}
\usage{
maxmin1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
named numeric vector with \code{max} and \code{min} of a random matrix
}
\description{
max() and min() of a matrix demo
}
