% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{mean1_}
\alias{mean1_}
\title{mean() column-wise, row-wise, and cube-wise mean demo}
\usage{
mean1_(X, Y)
}

\arguments{
\item{X}{numeric matrix (first input)}

\item{Y}{numeric matrix (second input used for cube)}
}

\value{
list with three numeric vectors of mean values
}

\description{
mean() column-wise, row-wise, and cube-wise mean demo
}

