% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{median1_}
\alias{median1_}
\title{median() column-wise and row-wise median demo}
\usage{
median1_(X, Y)
}

\arguments{
\item{X}{numeric matrix (first input)}

\item{Y}{numeric matrix (second input)}
}

\value{
list with three numeric vectors of median values
}

\description{
median() column-wise and row-wise median demo
}

