% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{misc1_}
\alias{misc1_}
\title{Miscellaneous element-wise math functions demo (exp, log, sqrt, round,
sign)}
\usage{
misc1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list of 6 matrices: original plus exp, log, sqrt, round, and sign
}
\description{
Miscellaneous element-wise math functions demo (exp, log, sqrt, round,
sign)
}
