% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{norm1_}
\alias{norm1_}
\title{norm() vector norm demo (L1, L2, inf, Frobenius)}
\usage{
norm1_(n)
}

\arguments{
\item{n}{integer scalar; vector length}
}

\value{
named numeric vector (5): five norm values
}

\description{
norm() vector norm demo (L1, L2, inf, Frobenius)
}

