% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{null1_}
\alias{null1_}
\title{null() null space computation demo}
\usage{
null1_(a)
}
\arguments{
\item{a}{numeric matrix (rows and first column zeroed)}
}
\value{
list with success flag and null space matrix
}
\description{
null() null space computation demo
}
