% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{ols_}
\alias{ols_}
\title{OLS regression via QR decomposition}
\usage{
ols_(x, y)
}
\arguments{
\item{x}{design matrix (N*P)}

\item{y}{response vector (N)}
}
\value{
numeric vector (P): OLS coefficient estimates
}
\description{
OLS regression via QR decomposition
}
