% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{ols_dbl}
\alias{ols_dbl}
\title{OLS coefficients as a vector}
\usage{
ols_dbl(y, x)
}
\arguments{
\item{y}{response matrix (N*1)}

\item{x}{design matrix (N*P)}
}
\value{
coefficient vector (P)
}
\description{
OLS coefficients as a vector
}
