% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{ols_qr_dbl}
\alias{ols_qr_dbl}
\title{OLS coefficients via QR decomposition as a vector}
\usage{
ols_qr_dbl(y, x, econ)
}

\arguments{
\item{y}{response matrix (N*1)}

\item{x}{design matrix (N*P)}

\item{econ}{logical scalar; if `TRUE` uses the economic QR decomposition}
}

\value{
coefficient vector (P)
}

\description{
OLS coefficients via QR decomposition as a vector
}

