% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{ols_qr_mat}
\alias{ols_qr_mat}
\title{OLS coefficients via QR decomposition as a matrix}
\usage{
ols_qr_mat(y, x, econ)
}
\arguments{
\item{y}{response matrix (N*1)}

\item{x}{design matrix (N*P)}

\item{econ}{logical scalar; if \code{TRUE} uses the economic QR decomposition}
}
\value{
coefficient matrix (P*1)
}
\description{
OLS coefficients via QR decomposition as a matrix
}
