% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{ols_simplified2_}
\alias{ols_simplified2_}
\title{OLS via QR returning a matrix using cpp4r::as_sexp()}
\usage{
ols_simplified2_(x, y)
}

\arguments{
\item{x}{design matrix (N*P)}

\item{y}{response vector (N)}
}

\value{
numeric matrix (P*1): OLS coefficient estimates
}

\description{
OLS via QR returning a matrix using cpp4r::as_sexp()
}

