% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{ols_traditional_}
\alias{ols_traditional_}
\title{OLS via QR using explicit as_doubles() conversion}
\usage{
ols_traditional_(x, y)
}

\arguments{
\item{x}{design matrix (N*P)}

\item{y}{response vector (N)}
}

\value{
numeric vector (P): OLS coefficient estimates
}

\description{
OLS via QR using explicit as_doubles() conversion
}

