% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{ones2_}
\alias{ones2_}
\title{ones() ones-filled container generation demo}
\usage{
ones2_(n)
}
\arguments{
\item{n}{integer scalar; container dimension}
}
\value{
numeric matrix (N*N): sum of various ones-filled vectors, matrices,
and a cube slice
}
\description{
ones() ones-filled container generation demo
}
