% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{orth1_}
\alias{orth1_}
\title{orth() orthonormal basis computation demo}
\usage{
orth1_(a)
}
\arguments{
\item{a}{numeric matrix}
}
\value{
list with success flag and orthonormal basis matrix
}
\description{
orth() orthonormal basis computation demo
}
