% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{polyval1_}
\alias{polyval1_}
\title{polyval() polynomial evaluation demo}
\usage{
polyval1_(n, m)
}

\arguments{
\item{n}{integer scalar; number of evaluation points}

\item{m}{integer scalar; polynomial degree}
}

\value{
numeric vector (N): polynomial values evaluated at N points
}

\description{
polyval() polynomial evaluation demo
}

