% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{prod1_}
\alias{prod1_}
\title{prod() column-wise and row-wise product demo}
\usage{
prod1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with column products and row products
}

\description{
prod() column-wise and row-wise product demo
}

