% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{qz1_}
\alias{qz1_}
\title{qz() generalized Schur (QZ) decomposition demo}
\usage{
qz1_(a, b, select)
}
\arguments{
\item{a}{numeric matrix (N*N)}

\item{b}{numeric matrix (N*N)}

\item{select}{string scalar; eigenvalue selection: \code{"lhp"}, \code{"rhp"}, \code{"iuc"},
\code{"ouc"}, or \code{"none"}}
}
\value{
list with success flag and AA, BB, Q, Z matrices
}
\description{
qz() generalized Schur (QZ) decomposition demo
}
