% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{randg3_}
\alias{randg3_}
\title{randi() integer random generation with optional distribution params demo}
\usage{
randg3_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): sum of integer random matrices with and
without distribution parameters
}
\description{
randi() integer random generation with optional distribution params demo
}
