% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{randn1_}
\alias{randn1_}
\title{Random normal matrix fill demo}
\usage{
randn1_(a)
}
\arguments{
\item{a}{numeric matrix}
}
\value{
numeric matrix: \code{a} plus two random normal matrices of the same size
}
\description{
Random normal matrix fill demo
}
