% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{random_cube_}
\alias{random_cube_}
\title{Create a random cube and return it as a list of matrices}
\usage{
random_cube_(n_rows, n_cols, n_slices)
}
\arguments{
\item{n_rows}{integer scalar; number of rows per slice}

\item{n_cols}{integer scalar; number of columns per slice}

\item{n_slices}{integer scalar; number of slices}
}
\value{
list of N_slices numeric matrices (each N_rows * N_cols)
}
\description{
Create a random cube and return it as a list of matrices
}
