% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{randu1_}
\alias{randu1_}
\title{Random uniform matrix fill demo}
\usage{
randu1_(a)
}

\arguments{
\item{a}{numeric matrix}
}

\value{
numeric matrix: `a` plus two random uniform matrices of the same size
}

\description{
Random uniform matrix fill demo
}

