% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{randu3_}
\alias{randu3_}
\title{randu() uniform random generation with optional distribution params demo}
\usage{
randu3_(n)
}
\arguments{
\item{n}{integer scalar; container dimension}
}
\value{
numeric matrix (N*N): sum of uniform random containers with and
without distribution parameters
}
\description{
randu() uniform random generation with optional distribution params demo
}
