% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{range1_}
\alias{range1_}
\title{range() column-wise and row-wise value range demo}
\usage{
range1_(X, Y)
}

\arguments{
\item{X}{numeric matrix (first input)}

\item{Y}{numeric matrix (second input)}
}

\value{
list with two numeric vectors of range values
}

\description{
range() column-wise and row-wise value range demo
}

