% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{rank1_}
\alias{rank1_}
\title{rank() matrix rank demo (two calling forms)}
\usage{
rank1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with scalar rank and a sub-list with scalar rank plus success
  flag
}

\description{
rank() matrix rank demo (two calling forms)
}

