% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{rcond1_}
\alias{rcond1_}
\title{rcond() reciprocal condition number demo}
\usage{
rcond1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric scalar: reciprocal condition number of a random matrix
}
\description{
rcond() reciprocal condition number demo
}
