% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{regspace1_}
\alias{regspace1_}
\title{regspace() regularly spaced vector demo}
\usage{
regspace1_(delta)
}

\arguments{
\item{delta}{numeric scalar; spacing between elements}
}

\value{
numeric vector: sum of a column and a row regspace vector
}

\description{
regspace() regularly spaced vector demo
}

