% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{reset1_}
\alias{reset1_}
\title{reset() matrix clearing demo}
\usage{
reset1_(n)
}

\arguments{
\item{n}{integer scalar; initial matrix dimension}
}

\value{
named integer vector with `n_rows` and `n_cols` after reset (both 0)
}

\description{
reset() matrix clearing demo
}

