% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{reshape1_}
\alias{reshape1_}
\title{reshape() dimension reordering demo}
\usage{
reshape1_(n)
}

\arguments{
\item{n}{integer scalar; base dimension}
}

\value{
numeric matrix ((N-1)*(N+1)) reshaped from an (N+1)*(N-1) matrix
}

\description{
reshape() dimension reordering demo
}

