% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{reshape2_}
\alias{reshape2_}
\title{reshape() functional form demo}
\usage{
reshape2_(n)
}

\arguments{
\item{n}{integer scalar; base matrix dimension}
}

\value{
list with three matrices: original and two reshaped versions
}

\description{
reshape() functional form demo
}

