% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{resize1_}
\alias{resize1_}
\title{resize() dimension change demo}
\usage{
resize1_(n)
}

\arguments{
\item{n}{integer scalar; base dimension}
}

\value{
numeric matrix ((N-1)*(N+1)) resized from an (N+1)*(N-1) matrix
}

\description{
resize() dimension change demo
}

