% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{resize2_}
\alias{resize2_}
\title{resize() functional form demo}
\usage{
resize2_(n)
}
\arguments{
\item{n}{integer scalar; base matrix dimension}
}
\value{
list with three matrices: original and two resized versions
}
\description{
resize() functional form demo
}
