% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{reverse1_}
\alias{reverse1_}
\title{reverse() row-wise and column-wise reversal demo}
\usage{
reverse1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with original, row-reversed, and column-reversed matrix
}

\description{
reverse() row-wise and column-wise reversal demo
}

