% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{row1_}
\alias{row1_}
\title{Row vector extraction demo}
\usage{
row1_(x, y)
}

\arguments{
\item{x}{numeric vector (length 10)}

\item{y}{numeric vector (length 10)}
}

\value{
numeric vector: row 6 of a random 10*10 matrix plus `y` plus `x`
}

\description{
Row vector extraction demo
}

