% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{row_as_mat1_}
\alias{row_as_mat1_}
\title{row_as_mat() cube row extraction demo}
\usage{
row_as_mat1_(n)
}
\arguments{
\item{n}{integer scalar; cube row dimension}
}
\value{
list with 4 slice matrices and the extracted \code{row_as_mat} matrix
}
\description{
row_as_mat() cube row extraction demo
}
