% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{saveload1_}
\alias{saveload1_}
\title{Matrix save and load demo (binary, text, CSV formats)}
\usage{
saveload1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
integer scalar: 0 on success
}
\description{
Matrix save and load demo (binary, text, CSV formats)
}
