% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{schur1_}
\alias{schur1_}
\title{schur() Schur decomposition demo}
\usage{
schur1_(x)
}

\arguments{
\item{x}{numeric matrix (N*N)}
}

\value{
list with success flag, unitary matrix U, and quasi-upper-triangular
  matrix S
}

\description{
schur() Schur decomposition demo
}

