% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{set_imag1_}
\alias{set_imag1_}
\title{set_imag() imaginary-part assignment demo}
\usage{
set_imag1_(n)
}

\arguments{
\item{n}{integer scalar; base matrix dimension}
}

\value{
complex matrix ((N+1)*(N-1)) with imaginary part set from a random matrix
}

\description{
set_imag() imaginary-part assignment demo
}

