% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{shed_columns1_}
\alias{shed_columns1_}
\title{shed_col() / shed_cols() column removal demo}
\usage{
shed_columns1_(n)
}

\arguments{
\item{n}{integer scalar; initial number of rows}
}

\value{
numeric matrix after removing 4 columns from an N*(5N) matrix
}

\description{
shed_col() / shed_cols() column removal demo
}

