% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{shed_rows1_}
\alias{shed_rows1_}
\title{shed_row() / shed_rows() row removal demo}
\usage{
shed_rows1_(n)
}

\arguments{
\item{n}{integer scalar; initial number of columns}
}

\value{
numeric matrix after removing 4 rows from a (5N)*N matrix
}

\description{
shed_row() / shed_rows() row removal demo
}

