% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{shuffle1_}
\alias{shuffle1_}
\title{shuffle() random row shuffling demo}
\usage{
shuffle1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with original and shuffled matrix
}
\description{
shuffle() random row shuffling demo
}
