% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{size1_}
\alias{size1_}
\title{size() object size operations demo}
\usage{
size1_(n)
}

\arguments{
\item{n}{integer scalar; base matrix dimension}
}

\value{
list of 7 matrices illustrating size() in construction, assignment,
  subview access, addition, and multiplication
}

\description{
size() object size operations demo
}

