% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{solve1_}
\alias{solve1_}
\title{solve() linear system solution demo}
\usage{
solve1_(a, b)
}

\arguments{
\item{a}{numeric matrix (N*N)}

\item{b}{numeric matrix (N*M)}
}

\value{
numeric matrix (N*M): solution X to A*X = B
}

\description{
solve() linear system solution demo
}

