% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{spdiags1_}
\alias{spdiags1_}
\title{spdiags() sparse banded matrix construction demo}
\usage{
spdiags1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): dense form of a lower triangular sparse banded
matrix
}
\description{
spdiags() sparse banded matrix construction demo
}
