% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{speye1_}
\alias{speye1_}
\title{speye() sparse identity matrix generation demo}
\usage{
speye1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): dense form of a sparse identity matrix
}

\description{
speye() sparse identity matrix generation demo
}

