% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{spones1_}
\alias{spones1_}
\title{spones() sparse nonzero-to-ones conversion demo}
\usage{
spones1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): dense form of a sparse matrix with nonzeros
  replaced by ones
}

\description{
spones() sparse nonzero-to-ones conversion demo
}

