% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sprandn1_}
\alias{sprandn1_}
\title{sprandn() random sparse normal matrix generation demo}
\usage{
sprandn1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): dense form of a random normal sparse matrix
}

\description{
sprandn() random sparse normal matrix generation demo
}

