% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{sqrtmat_sympd1_}
\alias{sqrtmat_sympd1_}
\title{sqrtmat_sympd() symmetric positive definite matrix square root demo}
\usage{
sqrtmat_sympd1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): real square root of a sym-pos-def matrix
}
\description{
sqrtmat_sympd() symmetric positive definite matrix square root demo
}
