% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{stddev1_}
\alias{stddev1_}
\title{stddev() standard deviation demo (sample and population)}
\usage{
stddev1_(X, Y)
}
\arguments{
\item{X}{numeric matrix (first input)}

\item{Y}{numeric matrix (second input)}
}
\value{
list with three numeric vectors of standard deviation values
}
\description{
stddev() standard deviation demo (sample and population)
}
