% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{subview1_}
\alias{subview1_}
\title{Submatrix and conditional element access demo}
\usage{
subview1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension (>= 5)}
}

\value{
numeric matrix (5*5) after subview assignments and element modifications
}

\description{
Submatrix and conditional element access demo
}

