% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{sum1_}
\alias{sum1_}
\title{sum() column-wise and row-wise summation demo}
\usage{
sum1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with row sums, column sums, and overall sum
}
\description{
sum() column-wise and row-wise summation demo
}
