% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sum2_}
\alias{sum2_}
\title{sum() column-wise and row-wise summation alternative form demo}
\usage{
sum2_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with row sums, column sums, and overall sum
}

\description{
sum() column-wise and row-wise summation alternative form demo
}

