% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{svd1_}
\alias{svd1_}
\title{svd() singular value decomposition demo}
\usage{
svd1_(x)
}

\arguments{
\item{x}{numeric matrix (M*N)}
}

\value{
list with success flag, U matrix, singular value vector, and V matrix
}

\description{
svd() singular value decomposition demo
}

