% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{swap1_}
\alias{swap1_}
\title{swap() matrix swap demo}
\usage{
swap1_(n)
}

\arguments{
\item{n}{integer scalar; dimension}
}

\value{
numeric matrix: result is the original `B` after swapping contents
  with `A`
}

\description{
swap() matrix swap demo
}

