% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{test_dgCMatrix_to_SpMat}
\alias{test_dgCMatrix_to_SpMat}
\title{Convert a dgCMatrix to SpMat\if{html}{\out{<double>}} and back}
\usage{
test_dgCMatrix_to_SpMat(x)
}
\arguments{
\item{x}{sparse matrix in dgCMatrix format}
}
\value{
dgCMatrix: round-trip result (values unchanged)
}
\description{
Convert a dgCMatrix to SpMat\if{html}{\out{<double>}} and back
}
