% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{test_sparse_dims}
\alias{test_sparse_dims}
\title{Return dimensions of a sparse matrix}
\usage{
test_sparse_dims(x)
}

\arguments{
\item{x}{sparse matrix (any format accepted by `as_SpMat()`)}
}

\value{
integer vector (2): number of rows and columns
}

\description{
Return dimensions of a sparse matrix
}

