% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{test_sparse_to_dgCMatrix}
\alias{test_sparse_to_dgCMatrix}
\title{Convert any sparse Matrix to dgCMatrix via SpMat<double>}
\usage{
test_sparse_to_dgCMatrix(x)
}

\arguments{
\item{x}{sparse matrix (any CsparseMatrix, RsparseMatrix, or TsparseMatrix)}
}

\value{
dgCMatrix: round-trip conversion result
}

\description{
Convert any sparse Matrix to dgCMatrix via SpMat<double>
}

