% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{toeplitz1_}
\alias{toeplitz1_}
\title{toeplitz() and circ_toeplitz() matrix generation demo}
\usage{
toeplitz1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): sum of a Toeplitz and a circulant Toeplitz
  matrix
}

\description{
toeplitz() and circ_toeplitz() matrix generation demo
}

