% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{transpose1_}
\alias{transpose1_}
\title{Matrix transpose demo}
\usage{
transpose1_(n)
}
\arguments{
\item{n}{integer scalar; matrix row dimension}
}
\value{
numeric matrix ((N+1)\emph{N): transpose of an N}(N+1) random matrix
}
\description{
Matrix transpose demo
}
