% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{transpose2_}
\alias{transpose2_}
\title{Sparse matrix transpose demo}
\usage{
transpose2_(n)
}
\arguments{
\item{n}{integer scalar; matrix row dimension}
}
\value{
numeric matrix ((N+1)\emph{N): transpose of a random sparse N}(N+1) matrix
}
\description{
Sparse matrix transpose demo
}
