% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{trapz1_}
\alias{trapz1_}
\title{trapz() trapezoidal numerical integration demo}
\usage{
trapz1_(n)
}

\arguments{
\item{n}{integer scalar; number of integration points}
}

\value{
numeric matrix (1*1): integral of sin(x) over 0 to pi
}

\description{
trapz() trapezoidal numerical integration demo
}

