% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{trig1_}
\alias{trig1_}
\title{Trigonometric element-wise functions demo (cos, sin, tan, atan2, hypot)}
\usage{
trig1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list of 6 matrices: original plus cos, sin, tan, atan2, and hypot
}

\description{
Trigonometric element-wise functions demo (cos, sin, tan, atan2, hypot)
}

