% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{typedef_Col_double}
\alias{typedef_Col_double}
\title{Round-trip conversion via Col\if{html}{\out{<double>}}}
\usage{
typedef_Col_double(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector identical to input
}
\description{
Round-trip conversion via Col\if{html}{\out{<double>}}
}
