% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{typedef_dblmat_exchangeability}
\alias{typedef_dblmat_exchangeability}
\title{Dense matrix typedef exchangeability}
\usage{
typedef_dblmat_exchangeability(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
list of 8 numeric matrices, each produced via a different Mat alias
}
\description{
Dense matrix typedef exchangeability
}
