% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{typedef_Mat_int}
\alias{typedef_Mat_int}
\title{Round-trip conversion via Mat<int>}
\usage{
typedef_Mat_int(x)
}

\arguments{
\item{x}{integer matrix}
}

\value{
integer matrix identical to input
}

\description{
Round-trip conversion via Mat<int>
}

