% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{typedef_spmat_exchangeability}
\alias{typedef_spmat_exchangeability}
\title{Sparse matrix typedef exchangeability (double)}
\usage{
typedef_spmat_exchangeability(x)
}

\arguments{
\item{x}{numeric matrix}
}

\value{
list of 3 numeric matrices, each produced via a different SpMat alias
}

\description{
Sparse matrix typedef exchangeability (double)
}

