% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{typedef_SpMat_int}
\alias{typedef_SpMat_int}
\title{Round-trip conversion via SpMat<int>}
\usage{
typedef_SpMat_int(x)
}

\arguments{
\item{x}{integer matrix}
}

\value{
integer matrix identical to input
}

\description{
Round-trip conversion via SpMat<int>
}

