% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{typedef_uvec}
\alias{typedef_uvec}
\title{Round-trip conversion via uvec (unsigned integer column vector)}
\usage{
typedef_uvec(x)
}
\arguments{
\item{x}{integer vector}
}
\value{
integer vector identical to input
}
\description{
Round-trip conversion via uvec (unsigned integer column vector)
}
