% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{unique1_}
\alias{unique1_}
\title{unique() sorted unique element extraction demo}
\usage{
unique1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric vector: unique elements of a random matrix
}

\description{
unique() sorted unique element extraction demo
}

