% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{uvec_example_}
\alias{uvec_example_}
\title{Unsigned integer matrix increment using as_umat() and cpp4r::as_sexp()}
\usage{
uvec_example_(x)
}
\arguments{
\item{x}{integer matrix}
}
\value{
integer matrix: input matrix with all elements incremented by 1
}
\description{
Unsigned integer matrix increment using as_umat() and cpp4r::as_sexp()
}
