% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{vec_demo_}
\alias{vec_demo_}
\title{R-to-Armadillo vector conversion demo (double and unsigned int)}
\usage{
vec_demo_(y, yi)
}

\arguments{
\item{y}{numeric vector}

\item{yi}{integer vector}
}

\value{
numeric vector: element-wise square of the input
}

\description{
R-to-Armadillo vector conversion demo (double and unsigned int)
}

