% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{vecnorm1_}
\alias{vecnorm1_}
\title{vecnorm() column-wise and row-wise vector norm demo}
\usage{
vecnorm1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with column L2 norms and row infinity norms
}
\description{
vecnorm() column-wise and row-wise vector norm demo
}
