% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{vectorise1_}
\alias{vectorise1_}
\title{vectorise() matrix-to-column-vector stacking demo}
\usage{
vectorise1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric vector (N^2): all elements of an N*N matrix in column-major
  order
}

\description{
vectorise() matrix-to-column-vector stacking demo
}

