% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{wishrnd1_}
\alias{wishrnd1_}
\title{wishrnd() Wishart random matrix sampling demo}
\usage{
wishrnd1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): sample from the Wishart distribution
}

\description{
wishrnd() Wishart random matrix sampling demo
}

