% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{zeros2_}
\alias{zeros2_}
\title{zeros() zeros-filled container generation demo}
\usage{
zeros2_(n)
}

\arguments{
\item{n}{integer scalar; container dimension}
}

\value{
numeric matrix (N*N): sum of various zeros-filled vectors, matrices,
  and a cube slice
}

\description{
zeros() zeros-filled container generation demo
}

