% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{access1_}
\alias{access1_}
\title{Element access and nested-loop mutation demo}
\usage{
access1_(a)
}

\arguments{
\item{a}{numeric matrix}
}

\value{
numeric matrix with each element increased by element (1,1) of the
  original matrix plus element 2 of a random vector
}

\description{
Element access and nested-loop mutation demo
}

