% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{accu1_}
\alias{accu1_}
\title{accu() element accumulation and multiply-accumulate demo}
\usage{
accu1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric scalar: sum of all elements of A plus sum of A element-wise
  multiplied by B
}

\description{
accu() element accumulation and multiply-accumulate demo
}

