% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{any1_}
\alias{any1_}
\title{any() any-element-nonzero check demo}
\usage{
any1_(n)
}

\arguments{
\item{n}{integer scalar; container dimension}
}

\value{
logical vector of four any() check results
}

\description{
any() any-element-nonzero check demo
}

