% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{approx_equal1_}
\alias{approx_equal1_}
\title{approx_equal() approximate equality check demo}
\usage{
approx_equal1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
logical scalar: \code{TRUE} if all three approximate equality checks pass
}
\description{
approx_equal() approximate equality check demo
}
