% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{arg1_}
\alias{arg1_}
\title{arg() complex argument (phase angle) demo}
\usage{
arg1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): element-wise phase angles of a complex matrix
}

\description{
arg() complex argument (phase angle) demo
}

