% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo4rtest-package.R
\docType{package}
\name{armadillo4rtest-package}
\alias{armadillo4rtest}
\alias{armadillo4rtest-package}
\title{armadillo4rtest: Testing 'armadillo4r'}
\description{
To check armadillo4r correctness. The headers are vendored to check which lines of the C++ code are covered by the unit tests.
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})
}

}
\keyword{internal}
