% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo4rtest-package.R
\docType{package}
\name{armadillo4rtest-package}
\alias{armadillo4rtest}
\alias{armadillo4rtest-package}
\title{armadillo4rtest: Testing armadillo4r}
\description{
Testing armadillo4r. I run this within a tiny package to test the compiled code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/armadillo4r}
  \item Report bugs at \url{https://github.com/pachadotdev/armadillo4r/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
