% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{as_col1_}
\alias{as_col1_}
\title{as_col() matrix-to-column-vector conversion demo}
\usage{
as_col1_(n)
}
\arguments{
\item{n}{integer scalar; matrix row dimension}
}
\value{
numeric vector: all elements of an N*(N+1) matrix stacked
column-major
}
\description{
as_col() matrix-to-column-vector conversion demo
}
