% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{as_row1_}
\alias{as_row1_}
\title{as_row() matrix-to-row-vector conversion demo}
\usage{
as_row1_(n)
}
\arguments{
\item{n}{integer scalar; matrix row dimension}
}
\value{
numeric vector: all elements of an N*(N+1) matrix stacked row-major
}
\description{
as_row() matrix-to-row-vector conversion demo
}
