% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{as_sexp_demo_}
\alias{as_sexp_demo_}
\title{Sort a vector using cpp4r::as_sexp() for the return value}
\usage{
as_sexp_demo_(y)
}
\arguments{
\item{y}{numeric vector}
}
\value{
numeric vector: sorted in ascending order
}
\description{
Sort a vector using cpp4r::as_sexp() for the return value
}
