% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{attr1_}
\alias{attr1_}
\title{Matrix dimension attributes}
\usage{
attr1_(a)
}

\arguments{
\item{a}{numeric matrix}
}

\value{
named integer vector with elements `n_rows`, `n_cols`, and `n_elem`
}

\description{
Matrix dimension attributes
}

