% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{build_field_}
\alias{build_field_}
\title{Build a field of random vectors of specified sizes}
\usage{
build_field_(sizes)
}
\arguments{
\item{sizes}{integer vector of vector lengths}
}
\value{
list of numeric vectors, each filled with standard normal values
}
\description{
Build a field of random vectors of specified sizes
}
