% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{capm}
\alias{capm}
\title{CAPM model}
\usage{
capm(r, m, f)
}
\arguments{
\item{r}{returns matrix (N*1)}

\item{m}{market returns matrix (N*1)}

\item{f}{risk-free rate (scalar)}
}
\value{
expected returns matrix (N*1)
}
\description{
CAPM model
}
