% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{chi2rnd1_}
\alias{chi2rnd1_}
\title{chi2rnd() chi-squared random sampling demo}
\usage{
chi2rnd1_(n, m)
}

\arguments{
\item{n}{integer scalar; number of rows}

\item{m}{integer scalar; number of columns}
}

\value{
list with two numeric matrices of chi-squared samples
}

\description{
chi2rnd() chi-squared random sampling demo
}

