% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{chol1_}
\alias{chol1_}
\title{chol() Cholesky decomposition with pivoting demo}
\usage{
chol1_(x, layout, output)
}

\arguments{
\item{x}{numeric positive semi-definite matrix (N*N)}

\item{layout}{string scalar; `"upper"` or `"lower"`}

\item{output}{string scalar; `"matrix"` for pivot matrix, `"vector"` for
  pivot vector}
}

\value{
list with success flag and triangular factor matrix
}

\description{
chol() Cholesky decomposition with pivoting demo
}

