% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{chol_mat}
\alias{chol_mat}
\title{Cholesky decomposition}
\usage{
chol_mat(x, type)
}

\arguments{
\item{x}{symmetric positive definite numeric matrix (N*N)}

\item{type}{decomposition type: `"upper"` for upper triangular or `"lower"`
  for lower triangular}
}

\value{
triangular factor matrix (N*N)
}

\description{
Cholesky decomposition
}

