% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{clamp2_}
\alias{clamp2_}
\title{clamp() functional form demo}
\usage{
clamp2_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): sum of three clamped versions of a random
matrix
}
\description{
clamp() functional form demo
}
