% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{col_as_mat1_}
\alias{col_as_mat1_}
\title{col_as_mat() cube column extraction demo}
\usage{
col_as_mat1_(n)
}
\arguments{
\item{n}{integer scalar; cube row dimension}
}
\value{
list with 4 slice matrices and the extracted \code{col_as_mat} matrix
}
\description{
col_as_mat() cube column extraction demo
}
