% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{column1_}
\alias{column1_}
\title{Column vector extraction demo}
\usage{
column1_(x, y)
}
\arguments{
\item{x}{numeric vector (length 10)}

\item{y}{numeric vector (length 10)}
}
\value{
numeric vector: column 6 of a random 10*10 matrix plus \code{y} plus \code{x}
}
\description{
Column vector extraction demo
}
