% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{compatibility1_}
\alias{compatibility1_}
\title{front() and back() compatibility demo}
\usage{
compatibility1_(n)
}
\arguments{
\item{n}{integer scalar; vector length}
}
\value{
named numeric vector with \code{front} and \code{back} elements
}
\description{
front() and back() compatibility demo
}
