% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{conv1_}
\alias{conv1_}
\title{conv() 1D convolution demo (full and same modes)}
\usage{
conv1_(x, y)
}

\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}

\value{
list with full-length and same-length convolution results
}

\description{
conv() 1D convolution demo (full and same modes)
}

