% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cross1_}
\alias{cross1_}
\title{cross() cross product demo}
\usage{
cross1_(n)
}
\arguments{
\item{n}{integer scalar; vector length (must be 3)}
}
\value{
numeric vector (3): cross product of two random vectors
}
\description{
cross() cross product demo
}
