% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cube1_}
\alias{cube1_}
\title{Cube (3-D array) construction and slice arithmetic demo}
\usage{
cube1_(a, b)
}
\arguments{
\item{a}{numeric matrix (first slice)}

\item{b}{numeric matrix (second slice)}
}
\value{
numeric matrix: sum of first slice of 2*cube plus second slice of
element-wise cube square
}
\description{
Cube (3-D array) construction and slice arithmetic demo
}
