% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{cube_col_means_}
\alias{cube_col_means_}
\title{Convert a list of matrices to a cube and compute column means per slice}
\usage{
cube_col_means_(slices)
}

\arguments{
\item{slices}{list of numeric matrices (all same dimensions)}
}

\value{
numeric matrix (N_slices * N_cols): column means for each slice
}

\description{
Convert a list of matrices to a cube and compute column means per slice
}

