% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{cumprod1_}
\alias{cumprod1_}
\title{cumprod() cumulative product along different dimensions demo}
\usage{
cumprod1_(n)
}

\arguments{
\item{n}{integer scalar; container dimension}
}

\value{
numeric vector (3): accumulated products of column-wise, row-wise,
  and vector cumulative products
}

\description{
cumprod() cumulative product along different dimensions demo
}

