% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{cumsum1_}
\alias{cumsum1_}
\title{cumsum() cumulative sum along different dimensions demo}
\usage{
cumsum1_(n)
}

\arguments{
\item{n}{integer scalar; container dimension}
}

\value{
numeric vector (3): accumulated sums of column-wise, row-wise, and
  vector cumulative sums
}

\description{
cumsum() cumulative sum along different dimensions demo
}

