% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{diagmat1_}
\alias{diagmat1_}
\title{diagmat() diagonal matrix construction demo}
\usage{
diagmat1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): sum of diagonal matrices built from a full
  matrix and a vector
}

\description{
diagmat() diagonal matrix construction demo
}

