% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{diagonal1_}
\alias{diagonal1_}
\title{Diagonal extraction and manipulation demo}
\usage{
diagonal1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric vector: sum of modified sparse and dense diagonals
}

\description{
Diagonal extraction and manipulation demo
}

