% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{diags1_}
\alias{diags1_}
\title{diags() banded matrix construction demo}
\usage{
diags1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): lower triangular banded matrix
}

\description{
diags() banded matrix construction demo
}

