% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{dot1_}
\alias{dot1_}
\title{dot(), cdot(), and norm_dot() inner product demo}
\usage{
dot1_(n)
}
\arguments{
\item{n}{integer scalar; vector length}
}
\value{
numeric vector (3): dot product, conjugate dot product, and
normalized dot product
}
\description{
dot(), cdot(), and norm_dot() inner product demo
}
