% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{each_slice1_}
\alias{each_slice1_}
\title{each_slice() broadcast demo}
\usage{
each_slice1_(n)
}

\arguments{
\item{n}{integer scalar; cube row dimension}
}

\value{
numeric matrix (N*(N+1)): sum of all slices after slice-wise
  broadcasts and lambda operations
}

\description{
each_slice() broadcast demo
}

