% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{eig_sym1_}
\alias{eig_sym1_}
\title{eig_sym() symmetric eigendecomposition demo (two forms)}
\usage{
eig_sym1_(x, method)
}

\arguments{
\item{x}{symmetric numeric matrix (N*N)}

\item{method}{string scalar; `"dc"` (divide-and-conquer) or `"std"`}
}

\value{
list with success flag, eigenvalue vector, and eigenvector matrix
}

\description{
eig_sym() symmetric eigendecomposition demo (two forms)
}

