% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{eig_sym2_}
\alias{eig_sym2_}
\title{eigs_sym() sparse symmetric eigendecomposition demo}
\usage{
eig_sym2_(x, method, k)
}
\arguments{
\item{x}{numeric matrix (used to form a sparse sym-pos-def matrix)}

\item{method}{string scalar; \code{"lm"} (largest magnitude) or \code{"sm"}}

\item{k}{integer scalar; number of eigenvalues to compute}
}
\value{
list with success flag, eigenvalue vector, and eigenvector matrix
}
\description{
eigs_sym() sparse symmetric eigendecomposition demo
}
