% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{eigen_gen_mat}
\alias{eigen_gen_mat}
\title{Eigenvalues of a general matrix (real and imaginary parts)}
\usage{
eigen_gen_mat(x)
}
\arguments{
\item{x}{numeric matrix (N*N)}
}
\value{
list with elements \code{real} and \code{imag} (each N*1 matrix)
}
\description{
Eigenvalues of a general matrix (real and imaginary parts)
}
