% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{eigen_sym_dbl}
\alias{eigen_sym_dbl}
\title{Eigenvalues of a symmetric matrix as a vector}
\usage{
eigen_sym_dbl(x)
}
\arguments{
\item{x}{symmetric numeric matrix (N*N)}
}
\value{
eigenvalue vector (N)
}
\description{
Eigenvalues of a symmetric matrix as a vector
}
