% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{eigen_sym_mat}
\alias{eigen_sym_mat}
\title{Eigenvalues of a symmetric matrix as a matrix}
\usage{
eigen_sym_mat(x)
}

\arguments{
\item{x}{symmetric numeric matrix (N*N)}
}

\value{
eigenvalue matrix (N*1)
}

\description{
Eigenvalues of a symmetric matrix as a matrix
}

