% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{eps1_}
\alias{eps1_}
\title{eps() machine epsilon per element demo}
\usage{
eps1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): element-wise machine epsilon of a random matrix
}

\description{
eps() machine epsilon per element demo
}

