% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{expmat1_}
\alias{expmat1_}
\title{expmat() matrix exponential demo}
\usage{
expmat1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): matrix exponential of a random matrix
}
\description{
expmat() matrix exponential demo
}
