% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{expmat_sym1_}
\alias{expmat_sym1_}
\title{expmat_sym() symmetric matrix exponential demo}
\usage{
expmat_sym1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): matrix exponential of a symmetrized random
matrix
}
\description{
expmat_sym() symmetric matrix exponential demo
}
