% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{eye1_}
\alias{eye1_}
\title{Identity matrix construction demo}
\usage{
eye1_(a)
}
\arguments{
\item{a}{numeric matrix}
}
\value{
numeric matrix: sum of three identity matrices of the same size
}
\description{
Identity matrix construction demo
}
