% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{eye2_}
\alias{eye2_}
\title{eye() identity matrix generation demo}
\usage{
eye2_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): sum of three identity matrices of different types
}
\description{
eye() identity matrix generation demo
}
