% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{fft2_}
\alias{fft2_}
\title{fft2() and ifft2() 2D fast Fourier transform demo}
\usage{
fft2_(x)
}

\arguments{
\item{x}{numeric matrix}
}

\value{
list with 2D FFT result (real/imag) and inverse 2D FFT result
  (real/imag)
}

\description{
fft2() and ifft2() 2D fast Fourier transform demo
}

