% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{field_str_demo_}
\alias{field_str_demo_}
\title{Append "!" to each string}
\usage{
field_str_demo_(x)
}

\arguments{
\item{x}{character vector}
}

\value{
character vector with "!" appended to every element
}

\description{
Append "!" to each string
}

