% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{field_vec_demo_}
\alias{field_vec_demo_}
\title{Scale each vector in a list by 2 via field<vec>}
\usage{
field_vec_demo_(vecs)
}

\arguments{
\item{vecs}{list of numeric vectors}
}

\value{
list of numeric vectors, each scaled by 2
}

\description{
Scale each vector in a list by 2 via field<vec>
}

