% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{fill1_}
\alias{fill1_}
\title{fill::value fill demo}
\usage{
fill1_(a)
}
\arguments{
\item{a}{numeric matrix used to determine size}
}
\value{
numeric matrix: sum of matrices filled with 200, 100, zeros, and ones
}
\description{
fill::value fill demo
}
