% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{fmat_example_}
\alias{fmat_example_}
\title{Float matrix scaling using as_fmat() and cpp4r::as_sexp()}
\usage{
fmat_example_(x)
}

\arguments{
\item{x}{numeric matrix}
}

\value{
numeric matrix: input scaled by 2 (via single-precision arithmetic)
}

\description{
Float matrix scaling using as_fmat() and cpp4r::as_sexp()
}

