% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{gmm1_}
\alias{gmm1_}
\title{gmm_diag Gaussian mixture model fitting and inference demo}
\usage{
gmm1_(n, d)
}
\arguments{
\item{n}{integer scalar; number of data points}

\item{d}{integer scalar; data dimensionality}
}
\value{
list with status, means, likelihoods, assignments, and histograms
}
\description{
gmm_diag Gaussian mixture model fitting and inference demo
}
