% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{has_inf1_}
\alias{has_inf1_}
\title{has_inf() infinity check demo (dense, cube, sparse)}
\usage{
has_inf1_(n)
}

\arguments{
\item{n}{integer scalar; container dimension}
}

\value{
named logical vector of three has-infinity predicate results
}

\description{
has_inf() infinity check demo (dense, cube, sparse)
}

