% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{hess1_}
\alias{hess1_}
\title{hess() Hessenberg decomposition demo}
\usage{
hess1_(x)
}

\arguments{
\item{x}{numeric matrix (N*N)}
}

\value{
list with success flag and upper Hessenberg matrix
}

\description{
hess() Hessenberg decomposition demo
}

