% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{histc1_}
\alias{histc1_}
\title{histc() cumulative histogram demo}
\usage{
histc1_(n)
}
\arguments{
\item{n}{integer scalar; number of random values}
}
\value{
integer vector: bin counts using custom edges
}
\description{
histc() cumulative histogram demo
}
