% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{imag1_}
\alias{imag1_}
\title{imag() and real() complex part extraction demo}
\usage{
imag1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with \code{imag} and \code{real} parts of a random complex matrix
}
\description{
imag() and real() complex part extraction demo
}
