% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{imbue2_}
\alias{imbue2_}
\title{imbue fill via R's RNG demo}
\usage{
imbue2_(a)
}
\arguments{
\item{a}{numeric matrix}
}
\value{
numeric matrix: \code{a} plus a matrix filled via \code{unif_rand()}
}
\description{
imbue fill via R's RNG demo
}
