% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{implicit_conversion_int}
\alias{implicit_conversion_int}
\title{Round-trip R matrix via arma::imat (signed integer)}
\usage{
implicit_conversion_int(x)
}
\arguments{
\item{x}{integer or numeric matrix}
}
\value{
integer matrix
}
\description{
Round-trip R matrix via arma::imat (signed integer)
}
