% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{ind2sub1_}
\alias{ind2sub1_}
\title{ind2sub() linear-index-to-subscript conversion demo}
\usage{
ind2sub1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with subscript vectors for a single index, a set of indices,
and a cube index
}
\description{
ind2sub() linear-index-to-subscript conversion demo
}
