% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{index_min1_}
\alias{index_min1_}
\title{index_max() column/row-wise argmax demo}
\usage{
index_min1_(n)
}
\arguments{
\item{n}{integer scalar; matrix row dimension}
}
\value{
numeric vector (2): max value in column 1 and max value in row N
}
\description{
index_max() column/row-wise argmax demo
}
