% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{inplace_trans1_}
\alias{inplace_trans1_}
\title{inplace_trans() in-place transpose demo}
\usage{
inplace_trans1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): matrix after in-place transpose
}

\description{
inplace_trans() in-place transpose demo
}

