% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{insert_columns1_}
\alias{insert_columns1_}
\title{insert_cols() column insertion demo}
\usage{
insert_columns1_(n)
}

\arguments{
\item{n}{integer scalar; base matrix dimension}
}

\value{
numeric matrix (3N-1)*(3N-1) with columns inserted
}

\description{
insert_cols() column insertion demo
}

