% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{integer_cube_roundtrip_}
\alias{integer_cube_roundtrip_}
\title{Round-trip list of matrices via Cube<int>}
\usage{
integer_cube_roundtrip_(slices)
}

\arguments{
\item{slices}{list of integer matrices}
}

\value{
list of integer matrices identical to input
}

\description{
Round-trip list of matrices via Cube<int>
}

