% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{integer_transpose_}
\alias{integer_transpose_}
\title{Integer matrix transpose using as_imat() and cpp4r::as_sexp()}
\usage{
integer_transpose_(x)
}
\arguments{
\item{x}{integer matrix}
}
\value{
integer matrix: transpose of the input
}
\description{
Integer matrix transpose using as_imat() and cpp4r::as_sexp()
}
