% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{interp2_}
\alias{interp2_}
\title{interp2() 2D interpolation demo}
\usage{
interp2_(n)
}

\arguments{
\item{n}{integer scalar; grid dimension}
}

\value{
numeric matrix: bilinearly interpolated surface on a finer grid
}

\description{
interp2() 2D interpolation demo
}

