% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{intersect1_}
\alias{intersect1_}
\title{intersect() sorted-set intersection demo}
\usage{
intersect1_(n)
}
\arguments{
\item{n}{integer scalar; sequence length}
}
\value{
integer vector: intersection of \code{n:1} and \code{2:(n+1)}
}
\description{
intersect() sorted-set intersection demo
}
