% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{inv1_}
\alias{inv1_}
\title{inv() matrix inversion with approximate fallback demo}
\usage{
inv1_(a)
}

\arguments{
\item{a}{numeric square matrix (N*N)}
}

\value{
list with success flag and inverse matrix
}

\description{
inv() matrix inversion with approximate fallback demo
}

