% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{is_diagonal1_}
\alias{is_diagonal1_}
\title{is_diagmat() diagonal matrix check demo}
\usage{
is_diagonal1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
named logical vector of three diagonal-matrix predicate results
}

\description{
is_diagmat() diagonal matrix check demo
}

