% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{is_square1_}
\alias{is_square1_}
\title{is_square() square matrix check demo}
\usage{
is_square1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
named logical vector of three square-matrix predicate results
}
\description{
is_square() square matrix check demo
}
