% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{is_triangular1_}
\alias{is_triangular1_}
\title{is_trimatu() / is_trimatl() triangularity check demo}
\usage{
is_triangular1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
named logical vector of three triangularity predicate results
}
\description{
is_trimatu() / is_trimatl() triangularity check demo
}
