% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{iterators2_}
\alias{iterators2_}
\title{Element and slice iterator demo on a cube}
\usage{
iterators2_(n)
}
\arguments{
\item{n}{integer scalar; cube dimension}
}
\value{
numeric matrix (N*(N+1)): sum over all slices after iterating
}
\description{
Element and slice iterator demo on a cube
}
