% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{iterators4_}
\alias{iterators4_}
\title{Range-based for loop over a subview demo}
\usage{
iterators4_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N) with subview elements set to 123
}
\description{
Range-based for loop over a subview demo
}
