% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{iwishrnd1_}
\alias{iwishrnd1_}
\title{iwishrnd() inverse-Wishart random matrix sampling demo}
\usage{
iwishrnd1_(n, d)
}

\arguments{
\item{n}{integer scalar; matrix dimension}

\item{d}{numeric scalar; degrees of freedom}
}

\value{
numeric matrix (N*N): sample from the inverse-Wishart distribution
}

\description{
iwishrnd() inverse-Wishart random matrix sampling demo
}

