% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{kmeans1_}
\alias{kmeans1_}
\title{kmeans() k-means clustering demo}
\usage{
kmeans1_(n, d)
}

\arguments{
\item{n}{integer scalar; number of data points}

\item{d}{integer scalar; data dimensionality}
}

\value{
list with success flag and cluster means matrix
}

\description{
kmeans() k-means clustering demo
}

