% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{log_det_sympd1_}
\alias{log_det_sympd1_}
\title{log_det_sympd() log-determinant of symmetric positive definite matrix
  demo}
\usage{
log_det_sympd1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with scalar log-det and a sub-list with scalar log-det plus
  success flag
}

\description{
log_det_sympd() log-determinant of symmetric positive definite matrix
  demo
}

