% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{logmat1_}
\alias{logmat1_}
\title{logmat() matrix logarithm demo}
\usage{
logmat1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
complex matrix (N*N): matrix logarithm of a random matrix
}

\description{
logmat() matrix logarithm demo
}

