% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{lognormpdf1_}
\alias{lognormpdf1_}
\title{log_normpdf() log normal density function demo (multiple forms)}
\usage{
lognormpdf1_(n)
}

\arguments{
\item{n}{integer scalar; vector length}
}

\value{
list with 5 log-density results using different parameter forms
}

\description{
log_normpdf() log normal density function demo (multiple forms)
}

