% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{logspace1_}
\alias{logspace1_}
\title{logspace() logarithmically spaced vector demo}
\usage{
logspace1_(n)
}

\arguments{
\item{n}{integer scalar; number of points}
}

\value{
numeric vector (N): sum of a column and a row logspace vector
}

\description{
logspace() logarithmically spaced vector demo
}

