% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{lu1_}
\alias{lu1_}
\title{lu() LU decomposition with pivoting demo}
\usage{
lu1_(x)
}
\arguments{
\item{x}{numeric matrix (N*N)}
}
\value{
list with success flag and L, U, P matrices
}
\description{
lu() LU decomposition with pivoting demo
}
