% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{mat_demo_}
\alias{mat_demo_}
\title{R-to-Armadillo matrix conversion demo (double, int, float)}
\usage{
mat_demo_(x, xi)
}
\arguments{
\item{x}{numeric matrix}

\item{xi}{integer matrix}
}
\value{
numeric matrix: A * t(A) where A is the double matrix
}
\description{
R-to-Armadillo matrix conversion demo (double, int, float)
}
