% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{matrix1_}
\alias{matrix1_}
\title{Matrix arithmetic demo}
\usage{
matrix1_(a)
}
\arguments{
\item{a}{numeric matrix}
}
\value{
numeric matrix combining scalar addition, matrix addition and
multiplication, and element-wise multiplication
}
\description{
Matrix arithmetic demo
}
