% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{matrix2_}
\alias{matrix2_}
\title{Complex matrix construction demo}
\usage{
matrix2_(a)
}

\arguments{
\item{a}{numeric matrix used as real part}
}

\value{
complex matrix constructed from two real matrices
}

\description{
Complex matrix construction demo
}

