% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{matrix_add_}
\alias{matrix_add_}
\title{Matrix addition using cpp4r::as_sexp()}
\usage{
matrix_add_(a, b)
}
\arguments{
\item{a}{numeric matrix}

\item{b}{numeric matrix}
}
\value{
numeric matrix: element-wise sum of \code{a} and \code{b}
}
\description{
Matrix addition using cpp4r::as_sexp()
}
