% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{maxmin1_}
\alias{maxmin1_}
\title{max() and min() of a matrix demo}
\usage{
maxmin1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
named numeric vector with `max` and `min` of a random matrix
}

\description{
max() and min() of a matrix demo
}

