% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{memptr1_}
\alias{memptr1_}
\title{memptr() raw memory pointer access demo}
\usage{
memptr1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N) with each element increased by 123 plus the
corresponding element of a second random matrix
}
\description{
memptr() raw memory pointer access demo
}
