% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in armadillo4rtest-package.R
\name{mtcars_mat}
\alias{mtcars_mat}
\title{Mtcars dataset in matrix form}
\description{
Reshaped dataframe to test Armadillo linear algebra functions with
simple linear models of the form mpg_i = a + b cyl_i or
mpg_i = b cyl_i + c_1 cyl4_i + c_2 cyl6_i + c_3 cyl8_i
}

