% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{mvnrnd1_}
\alias{mvnrnd1_}
\title{mvnrnd() multivariate normal random sampling demo}
\usage{
mvnrnd1_(n, m)
}
\arguments{
\item{n}{integer scalar; dimension of the distribution}

\item{m}{integer scalar; number of samples}
}
\value{
numeric matrix (N*M): multivariate normal samples
}
\description{
mvnrnd() multivariate normal random sampling demo
}
