% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{nonzeros1_}
\alias{nonzeros1_}
\title{nonzeros() non-zero element extraction demo}
\usage{
nonzeros1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric vector: non-zero elements of a sparse-like matrix
}
\description{
nonzeros() non-zero element extraction demo
}
