% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{norm1_}
\alias{norm1_}
\title{norm() vector norm demo (L1, L2, inf, Frobenius)}
\usage{
norm1_(n)
}
\arguments{
\item{n}{integer scalar; vector length}
}
\value{
named numeric vector (5): five norm values
}
\description{
norm() vector norm demo (L1, L2, inf, Frobenius)
}
