% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{norm2est1_}
\alias{norm2est1_}
\title{norm2est() 2-norm estimation demo}
\usage{
norm2est1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric scalar: estimated 2-norm of a random matrix
}
\description{
norm2est() 2-norm estimation demo
}
