% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{normalise1_}
\alias{normalise1_}
\title{normalise() column-wise and row-wise normalization demo}
\usage{
normalise1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with column-normalized and row-normalized matrices
}
\description{
normalise() column-wise and row-wise normalization demo
}
