% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{normcdf1_}
\alias{normcdf1_}
\title{normcdf() normal cumulative distribution function demo (multiple forms)}
\usage{
normcdf1_(n)
}
\arguments{
\item{n}{integer scalar; vector length}
}
\value{
list with 5 CDF results using different parameter forms
}
\description{
normcdf() normal cumulative distribution function demo (multiple forms)
}
