% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{normpdf1_}
\alias{normpdf1_}
\title{normpdf() normal probability density function demo (multiple forms)}
\usage{
normpdf1_(n)
}

\arguments{
\item{n}{integer scalar; vector length}
}

\value{
list with 5 density results using different parameter forms
}

\description{
normpdf() normal probability density function demo (multiple forms)
}

