% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{ols_mat}
\alias{ols_mat}
\title{OLS coefficients as a matrix}
\usage{
ols_mat(y, x)
}
\arguments{
\item{y}{response matrix (N*1)}

\item{x}{design matrix (N*P)}
}
\value{
coefficient matrix (P*1)
}
\description{
OLS coefficients as a matrix
}
