% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{ols_simplified3_}
\alias{ols_simplified3_}
\title{OLS via QR using cpp4r::as_sexp() (form 3)}
\usage{
ols_simplified3_(x, y)
}

\arguments{
\item{x}{design matrix (N*P)}

\item{y}{response vector (N)}
}

\value{
numeric vector (P): OLS coefficient estimates
}

\description{
OLS via QR using cpp4r::as_sexp() (form 3)
}

