% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{ols_simplified4_}
\alias{ols_simplified4_}
\title{One-liner OLS using cpp4r::as_sexp() and as_col()}
\usage{
ols_simplified4_(x, y)
}
\arguments{
\item{x}{design matrix (N*P)}

\item{y}{response vector (N)}
}
\value{
numeric vector (P): OLS coefficient estimates
}
\description{
One-liner OLS using cpp4r::as_sexp() and as_col()
}
