% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{ols_simplified5_}
\alias{ols_simplified5_}
\title{One-liner OLS returning a matrix using cpp4r::as_sexp() and as_col()}
\usage{
ols_simplified5_(x, y)
}
\arguments{
\item{x}{design matrix (N*P)}

\item{y}{response vector (N)}
}
\value{
numeric matrix (P*1): OLS coefficient estimates
}
\description{
One-liner OLS returning a matrix using cpp4r::as_sexp() and as_col()
}
