% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{ols_simplified_}
\alias{ols_simplified_}
\title{OLS via QR using cpp4r::as_sexp() for vec-to-doubles conversion}
\usage{
ols_simplified_(x, y)
}

\arguments{
\item{x}{design matrix (N*P)}

\item{y}{response vector (N)}
}

\value{
numeric vector (P): OLS coefficient estimates
}

\description{
OLS via QR using cpp4r::as_sexp() for vec-to-doubles conversion
}

