% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{ones1_}
\alias{ones1_}
\title{Matrix ones fill demo}
\usage{
ones1_(a)
}

\arguments{
\item{a}{numeric matrix}
}

\value{
numeric matrix: sum of three ones-filled matrices of the same size
}

\description{
Matrix ones fill demo
}

