% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{pinv1_}
\alias{pinv1_}
\title{pinv() Moore-Penrose pseudoinverse demo}
\usage{
pinv1_(a)
}

\arguments{
\item{a}{numeric matrix}
}

\value{
list with pseudoinverse matrix
}

\description{
pinv() Moore-Penrose pseudoinverse demo
}

