% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{poisson_}
\alias{poisson_}
\title{Poisson regression via iteratively reweighted least squares}
\usage{
poisson_(x, y)
}

\arguments{
\item{x}{design matrix (N*P)}

\item{y}{response vector (N) of non-negative counts}
}

\value{
numeric vector (P): Poisson GLM coefficient estimates
}

\description{
Poisson regression via iteratively reweighted least squares
}

