% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{polyfit1_}
\alias{polyfit1_}
\title{polyfit() polynomial curve fitting demo}
\usage{
polyfit1_(n, m)
}
\arguments{
\item{n}{integer scalar; number of data points}

\item{m}{integer scalar; polynomial degree}
}
\value{
numeric vector (M+1): fitted polynomial coefficients
}
\description{
polyfit() polynomial curve fitting demo
}
