% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{pow1_}
\alias{pow1_}
\title{pow() element-wise power demo (scalar and matrix exponents)}
\usage{
pow1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with squared matrix and element-wise powered matrix
}

\description{
pow() element-wise power demo (scalar and matrix exponents)
}

