% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{powmat1_}
\alias{powmat1_}
\title{powmat() matrix power demo (two calling forms)}
\usage{
powmat1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with direct result and a sub-list with result plus success flag
}
\description{
powmat() matrix power demo (two calling forms)
}
