% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{qr1_}
\alias{qr1_}
\title{qr() QR decomposition demo}
\usage{
qr1_(x)
}

\arguments{
\item{x}{numeric matrix}
}

\value{
list with success flag, Q matrix, and R matrix
}

\description{
qr() QR decomposition demo
}

