% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{qr_econ1_}
\alias{qr_econ1_}
\title{qr_econ() economic QR decomposition demo}
\usage{
qr_econ1_(x)
}

\arguments{
\item{x}{numeric matrix (N*P, N >= P)}
}

\value{
list with success flag, thin Q matrix (N*P), and R matrix (P*P)
}

\description{
qr_econ() economic QR decomposition demo
}

