% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{quantile1_}
\alias{quantile1_}
\title{quantile() empirical quantile demo}
\usage{
quantile1_(n)
}
\arguments{
\item{n}{integer scalar; number of random values}
}
\value{
numeric vector (5): 0th, 25th, 50th, 75th, 100th percentiles
}
\description{
quantile() empirical quantile demo
}
