% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{randn1_}
\alias{randn1_}
\title{Random normal matrix fill demo}
\usage{
randn1_(a)
}

\arguments{
\item{a}{numeric matrix}
}

\value{
numeric matrix: `a` plus two random normal matrices of the same size
}

\description{
Random normal matrix fill demo
}

