% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{randn2_}
\alias{randn2_}
\title{Random normal matrix using R's RNG}
\usage{
randn2_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N) with standard normal random values
}

\description{
Random normal matrix using R's RNG
}

