% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{randn3_}
\alias{randn3_}
\title{randn() normal random generation with optional distribution params demo}
\usage{
randn3_(n)
}

\arguments{
\item{n}{integer scalar; container dimension}
}

\value{
numeric matrix (N*N): sum of normal random containers with and
  without distribution parameters
}

\description{
randn() normal random generation with optional distribution params demo
}

