% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{random_matrix_nxn}
\alias{random_matrix_nxn}
\title{Random integer matrix using R's RNG}
\usage{
random_matrix_nxn(n)
}

\arguments{
\item{n}{integer scalar; number of rows and columns}
}

\value{
integer matrix (N*N) with values in the range 0 to 9
}

\description{
Random integer matrix using R's RNG
}

