% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{randperm1_}
\alias{randperm1_}
\title{randperm() random permutation demo}
\usage{
randperm1_(n, m)
}
\arguments{
\item{n}{integer scalar; permutation size}

\item{m}{integer scalar; number of elements to draw}
}
\value{
integer vector: concatenation of a full and a partial permutation
}
\description{
randperm() random permutation demo
}
