% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{randu2_}
\alias{randu2_}
\title{Random uniform matrix using R's RNG}
\usage{
randu2_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N) with uniform random values
}

\description{
Random uniform matrix using R's RNG
}

