% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{range1_}
\alias{range1_}
\title{range() column-wise and row-wise value range demo}
\usage{
range1_(X, Y)
}
\arguments{
\item{X}{numeric matrix (first input)}

\item{Y}{numeric matrix (second input)}
}
\value{
list with two numeric vectors of range values
}
\description{
range() column-wise and row-wise value range demo
}
