% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{repelem1_}
\alias{repelem1_}
\title{repelem() element repetition demo}
\usage{
repelem1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with original and repeated (2x rows, 3x columns) matrix
}
\description{
repelem() element repetition demo
}
