% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{replace1_}
\alias{replace1_}
\title{replace() NaN-with-zero demo}
\usage{
replace1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N) with diagonal NaNs replaced by zero
}

\description{
replace() NaN-with-zero demo
}

