% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{repmat1_}
\alias{repmat1_}
\title{repmat() matrix tiling demo}
\usage{
repmat1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with original and tiled (2x rows, 3x columns) matrix
}
\description{
repmat() matrix tiling demo
}
