% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{reshape2_}
\alias{reshape2_}
\title{reshape() functional form demo}
\usage{
reshape2_(n)
}
\arguments{
\item{n}{integer scalar; base matrix dimension}
}
\value{
list with three matrices: original and two reshaped versions
}
\description{
reshape() functional form demo
}
