% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{roots1_}
\alias{roots1_}
\title{roots() polynomial root finding demo}
\usage{
roots1_(n)
}

\arguments{
\item{n}{integer scalar; polynomial degree}
}

\value{
complex vector (N-1): roots of a random degree-(N-1) polynomial
}

\description{
roots() polynomial root finding demo
}

