% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{saveload2_}
\alias{saveload2_}
\title{Field-of-matrices save and load demo (binary and PPM formats)}
\usage{
saveload2_(n, m)
}

\arguments{
\item{n}{integer scalar; matrix dimension within the field}

\item{m}{integer scalar; number of matrices in the field}
}

\value{
integer scalar: 0 on success
}

\description{
Field-of-matrices save and load demo (binary and PPM formats)
}

