% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{set_real1_}
\alias{set_real1_}
\title{set_real() real-part assignment demo}
\usage{
set_real1_(n)
}
\arguments{
\item{n}{integer scalar; base matrix dimension}
}
\value{
complex matrix ((N+1)*(N-1)) with real part set from a random matrix
}
\description{
set_real() real-part assignment demo
}
