% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{set_real2_}
\alias{set_real2_}
\title{cx_mat construction from two real matrices demo}
\usage{
set_real2_(n)
}
\arguments{
\item{n}{integer scalar; base matrix dimension}
}
\value{
complex matrix ((N-1)*(N+1)) constructed from two random real matrices
}
\description{
cx_mat construction from two real matrices demo
}
