% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{set_size1_}
\alias{set_size1_}
\title{set_size() memory allocation demo}
\usage{
set_size1_(n)
}
\arguments{
\item{n}{integer scalar; container dimension}
}
\value{
numeric vector: sum of first columns of two matrices plus a vector,
all of size N
}
\description{
set_size() memory allocation demo
}
