% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{shed_slices1_}
\alias{shed_slices1_}
\title{shed_slice() / shed_slices() cube slice removal demo}
\usage{
shed_slices1_(n)
}

\arguments{
\item{n}{integer scalar; cube dimension}
}

\value{
numeric matrix (N*N): sum over remaining slices after removing 4
}

\description{
shed_slice() / shed_slices() cube slice removal demo
}

