% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{shift1_}
\alias{shift1_}
\title{shift() element shifting demo (negative and positive)}
\usage{
shift1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with original, left-shifted, and right-shifted matrix
}

\description{
shift() element shifting demo (negative and positive)
}

