% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sort1_}
\alias{sort1_}
\title{sort() ascending/descending column and row sort demo}
\usage{
sort1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list of 5 matrices: original, col-asc, col-desc, row-asc, row-desc
}

\description{
sort() ascending/descending column and row sort demo
}

