% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sort_index1_}
\alias{sort_index1_}
\title{sort_index() ascending and descending sort index demo}
\usage{
sort_index1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with original matrix and two index vectors
}

\description{
sort_index() ascending and descending sort index demo
}

