% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{sprandu1_}
\alias{sprandu1_}
\title{sprandu() random sparse uniform matrix generation demo}
\usage{
sprandu1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): dense form of a random sparse matrix
}
\description{
sprandu() random sparse uniform matrix generation demo
}
