% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{spsolve1_}
\alias{spsolve1_}
\title{spsolve() sparse linear system solver demo}
\usage{
spsolve1_(a, b, method)
}
\arguments{
\item{a}{numeric matrix (used as a sparse matrix N*N)}

\item{b}{numeric vector (N)}

\item{method}{string scalar; \code{"superlu"} or \code{"lapack"}}
}
\value{
numeric vector (N): solution X to A*X = B
}
\description{
spsolve() sparse linear system solver demo
}
