% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{sqrtmat1_}
\alias{sqrtmat1_}
\title{sqrtmat() matrix square root demo (two calling forms)}
\usage{
sqrtmat1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
list with original matrix, two complex square roots, and success flag
}

\description{
sqrtmat() matrix square root demo (two calling forms)
}

