% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{sub2ind1_}
\alias{sub2ind1_}
\title{sub2ind() subscript-to-linear-index conversion demo}
\usage{
sub2ind1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
integer scalar: linear index of element (N-1, N-1)
}
\description{
sub2ind() subscript-to-linear-index conversion demo
}
