% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{svd_econ1_}
\alias{svd_econ1_}
\title{svd_econ() economic SVD demo}
\usage{
svd_econ1_(x)
}

\arguments{
\item{x}{numeric matrix (M*N)}
}

\value{
list with thin U matrix, singular value vector, and thin V matrix
}

\description{
svd_econ() economic SVD demo
}

