% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{svds1_}
\alias{svds1_}
\title{svds() sparse singular value decomposition demo}
\usage{
svds1_(x, k)
}

\arguments{
\item{x}{numeric matrix (used as a sparse matrix)}

\item{k}{integer scalar; number of singular values to compute}
}

\value{
list with success flag, singular value vector, U matrix, and V matrix
}

\description{
svds() sparse singular value decomposition demo
}

