% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{swap_columns1_}
\alias{swap_columns1_}
\title{swap_cols() column swapping demo}
\usage{
swap_columns1_(n)
}
\arguments{
\item{n}{integer scalar; number of rows}
}
\value{
numeric matrix (N*(5N)) after two column swaps
}
\description{
swap_cols() column swapping demo
}
