% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{swap_rows1_}
\alias{swap_rows1_}
\title{swap_rows() row swapping demo}
\usage{
swap_rows1_(n)
}
\arguments{
\item{n}{integer scalar; number of columns}
}
\value{
numeric matrix ((5N)*N) after two row swaps
}
\description{
swap_rows() row swapping demo
}
