% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{syl1_}
\alias{syl1_}
\title{sylvester() Sylvester equation solver (A\emph{X + X}B = C) demo}
\usage{
syl1_(a, b, c)
}
\arguments{
\item{a}{numeric matrix (N*N)}

\item{b}{numeric matrix (N*N)}

\item{c}{numeric matrix (N*N)}
}
\value{
numeric matrix (N*N): solution X
}
\description{
sylvester() Sylvester equation solver (A\emph{X + X}B = C) demo
}
