% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{symmatu1_}
\alias{symmatu1_}
\title{symmatu() upper-triangular symmetrization demo}
\usage{
symmatu1_(n)
}

\arguments{
\item{n}{integer scalar; matrix dimension}
}

\value{
numeric matrix (N*N): symmetric matrix from upper triangle of input
}

\description{
symmatu() upper-triangular symmetrization demo
}

