% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{test_sparse_to_dgCMatrix}
\alias{test_sparse_to_dgCMatrix}
\title{Convert any sparse Matrix to dgCMatrix via SpMat\if{html}{\out{<double>}}}
\usage{
test_sparse_to_dgCMatrix(x)
}
\arguments{
\item{x}{sparse matrix (any CsparseMatrix, RsparseMatrix, or TsparseMatrix)}
}
\value{
dgCMatrix: round-trip conversion result
}
\description{
Convert any sparse Matrix to dgCMatrix via SpMat\if{html}{\out{<double>}}
}
