% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{trace1_}
\alias{trace1_}
\title{trace() matrix trace demo}
\usage{
trace1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric scalar: trace of a random N*N matrix
}
\description{
trace() matrix trace demo
}
