% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{trans1_}
\alias{trans1_}
\title{trans() / .t() matrix transpose demo}
\usage{
trans1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
list with original matrix and its transpose
}
\description{
trans() / .t() matrix transpose demo
}
