% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{transform1_}
\alias{transform1_}
\title{transform() element-wise function application demo}
\usage{
transform1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N) with 122 added to every element
}
\description{
transform() element-wise function application demo
}
