% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{transpose1_}
\alias{transpose1_}
\title{Matrix transpose demo}
\usage{
transpose1_(n)
}

\arguments{
\item{n}{integer scalar; matrix row dimension}
}

\value{
numeric matrix ((N+1)*N): transpose of an N*(N+1) random matrix
}

\description{
Matrix transpose demo
}

