% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{trimatu1_}
\alias{trimatu1_}
\title{trimatu() upper triangular extraction demo}
\usage{
trimatu1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
numeric matrix (N*N): upper triangular part of a random matrix
}
\description{
trimatu() upper triangular extraction demo
}
