% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{trimatu_ind1_}
\alias{trimatu_ind1_}
\title{trimatu_ind() upper triangular element indices demo}
\usage{
trimatu_ind1_(n)
}
\arguments{
\item{n}{integer scalar; matrix dimension}
}
\value{
integer vector: linear indices of upper triangular elements
}
\description{
trimatu_ind() upper triangular element indices demo
}
