% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{typedef_Col_double}
\alias{typedef_Col_double}
\title{Round-trip conversion via Col<double>}
\usage{
typedef_Col_double(x)
}

\arguments{
\item{x}{numeric vector}
}

\value{
numeric vector identical to input
}

\description{
Round-trip conversion via Col<double>
}

