% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{typedef_Col_int}
\alias{typedef_Col_int}
\title{Round-trip conversion via Col\if{html}{\out{<int>}}}
\usage{
typedef_Col_int(x)
}
\arguments{
\item{x}{integer vector}
}
\value{
integer vector identical to input
}
\description{
Round-trip conversion via Col\if{html}{\out{<int>}}
}
