% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{typedef_Mat_double}
\alias{typedef_Mat_double}
\title{Round-trip conversion via Mat<double>}
\usage{
typedef_Mat_double(x)
}

\arguments{
\item{x}{numeric matrix}
}

\value{
numeric matrix identical to input
}

\description{
Round-trip conversion via Mat<double>
}

