% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{typedef_SpMat_double}
\alias{typedef_SpMat_double}
\title{Round-trip conversion via SpMat\if{html}{\out{<double>}}}
\usage{
typedef_SpMat_double(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
numeric matrix identical to input
}
\description{
Round-trip conversion via SpMat\if{html}{\out{<double>}}
}
