% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{typedef_SpMat_int}
\alias{typedef_SpMat_int}
\title{Round-trip conversion via SpMat\if{html}{\out{<int>}}}
\usage{
typedef_SpMat_int(x)
}
\arguments{
\item{x}{integer matrix}
}
\value{
integer matrix identical to input
}
\description{
Round-trip conversion via SpMat\if{html}{\out{<int>}}
}
