% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{typedef_spmat_int_exchangeability}
\alias{typedef_spmat_int_exchangeability}
\title{Sparse matrix typedef exchangeability (integer)}
\usage{
typedef_spmat_int_exchangeability(x)
}

\arguments{
\item{x}{integer matrix}
}

\value{
list of 3 integer matrices, each produced via a different integer SpMat alias
}

\description{
Sparse matrix typedef exchangeability (integer)
}

