% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{var1_}
\alias{var1_}
\title{var() variance demo (sample and population)}
\usage{
var1_(X, Y)
}

\arguments{
\item{X}{numeric matrix (first input)}

\item{Y}{numeric matrix (second input)}
}

\value{
list with three numeric vectors of variance values
}

\description{
var() variance demo (sample and population)
}

