% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{vec_out_demo_}
\alias{vec_out_demo_}
\title{Armadillo-to-R vector conversion demo: normalise()}
\usage{
vec_out_demo_(y)
}

\arguments{
\item{y}{numeric vector}
}

\value{
numeric vector: L2-normalized input vector
}

\description{
Armadillo-to-R vector conversion demo: normalise()
}

