% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{zeros1_}
\alias{zeros1_}
\title{Matrix zeros fill demo}
\usage{
zeros1_(a)
}

\arguments{
\item{a}{numeric matrix}
}

\value{
numeric matrix: sum of three zero-filled matrices of the same size
}

\description{
Matrix zeros fill demo
}

